feat(settings): full-page Mintlify redesign — sidebar-extension rail, carded sections, one-click API keys
Retire the settings modal for a real /settings/* page: a persistent left
rail that reads as an extension of the app sidebar, beside the active
tab. Adopt a Mintlify-style design language — a borderless left-gutter
section header paired with subtle cards on the right, plus full-width
sections for overview/stats. Every tab inherits it through the shared
SettingsSection primitive (props unchanged), so the look propagates
without per-tab rewrites.

Flagship: API-key and OAuth-client creation is now one-click — mint with
sensible defaults, reveal the secret inline with a copy button — plus an
optional "Custom…" name/scope path. No /new form or /created subpage; the
reveal lands inline via the in-page swapper (or a session-flash fallback
with JS off).

Shell
- New SettingsPage (rail + content) + settings_page.js (in-page tab swap
  + fragment-POST saves + toast), replacing the settings modal. The gear,
  cmdk, and shortcuts now navigate to /settings/<tab> via one global
  open-settings listener.
- renderSettingsTab renders the page on full loads, body fragments on
  swaps; resolves role flags so the rail gates per-tab correctly.

Design language
- SettingsSection: borderless two-column (gutter header | content), the
  content wrapped in a subtle card by default; new FullWidth (no gutter)
  and Bare (no card) variants, and a SettingsCard primitive.
- Rail mirrors the bb-sidebar-link language and sits flush against the app
  sidebar as a base-100 panel; grouped Workspace / Integrations / Access /
  Maintenance.
- Per-tab grids that placed two sections side-by-side are stacked; gutter
  titles top-align with their card.

Cleanup
- Remove API-key/OAuth subpages (routes, handlers, templ) and the dead
  settings_modal.*, settings_host, and mcp_settings.* trees.
- Backups restore warning → daisy alert.

Validated across tabs, dark mode, and mobile; go vet + go test ./... pass.
