+++ internal/admin/apikeys.go
@@ -7,7 +7,6 @@ import (
 	"strings"
 
 	"breadbox/internal/service"
-	"breadbox/internal/templates/components"
 	"breadbox/internal/templates/components/pages"
 	"breadbox/internal/timefmt"
 
@@ -73,53 +72,76 @@ func RevokeAPIKeyHandler(svc *service.Service) http.HandlerFunc {
 // AccessPageHandler serves GET /admin/settings/api-keys — combined API Keys + OAuth Clients page.
 func AccessPageHandler(svc *service.Service, sm *scs.SessionManager, tr *TemplateRenderer) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
-		keys, err := svc.ListAPIKeys(r.Context())
-		if err != nil {
-			http.Error(w, "Internal server error", http.StatusInternalServerError)
-			return
-		}
-		clients, err := svc.ListOAuthClients(r.Context())
-		if err != nil {
-			http.Error(w, "Internal server error", http.StatusInternalServerError)
-			return
-		}
-		// Split keys into active and revoked for cleaner display.
-		var activeKeys, revokedKeys []pages.AccessKeyRow
-		for _, k := range keys {
-			row := buildAccessKeyRow(k)
-			if k.RevokedAt != nil {
-				revokedKeys = append(revokedKeys, row)
-			} else {
-				activeKeys = append(activeKeys, row)
-			}
-		}
-		var activeClients, revokedClients []pages.AccessClientRow
-		for _, c := range clients {
-			row := buildAccessClientRow(c)
-			if c.RevokedAt != nil {
-				revokedClients = append(revokedClients, row)
-			} else {
-				activeClients = append(activeClients, row)
-			}
-		}
-		data := BaseTemplateData(r, sm, "api-keys","Access")
-		props := pages.AccessProps{
-			IsAdmin:        IsAdmin(sm, r),
-			CSRFToken:      GetCSRFToken(r),
-			ActiveKeys:     activeKeys,
-			RevokedKeys:    revokedKeys,
-			HasAnyKeys:     len(keys) > 0,
-			ActiveClients:  activeClients,
-			RevokedClients: revokedClients,
-			HasAnyClients:  len(clients) > 0,
-		}
-		renderAccess(w, r, sm, tr, data, props)
+		// On a cold/no-JS load the reveal arrives via session flash
+		// (set by a create that 303-redirected here). The JS create path
+		// never lands here — it renders the fragment directly.
+		renderAccessTab(svc, sm, tr, w, r, popAPIKeyReveal(r, sm), popOAuthClientReveal(r, sm))
 	}
 }
 
-// renderAccess wraps the Access tab body in the unified Settings shell.
-func renderAccess(w http.ResponseWriter, r *http.Request, sm *scs.SessionManager, tr *TemplateRenderer, data map[string]any, props pages.AccessProps) {
+// renderAccessTab lists keys + clients and renders the Access tab,
+// optionally with a one-time reveal block for a just-created key or
+// client. Shared by AccessPageHandler (GET) and the create handlers so a
+// fragment create can paint the reveal inline without a redirect.
+func renderAccessTab(svc *service.Service, sm *scs.SessionManager, tr *TemplateRenderer, w http.ResponseWriter, r *http.Request, keyReveal, clientReveal *pages.AccessReveal) {
+	keys, err := svc.ListAPIKeys(r.Context())
+	if err != nil {
+		http.Error(w, "Internal server error", http.StatusInternalServerError)
+		return
+	}
+	clients, err := svc.ListOAuthClients(r.Context())
+	if err != nil {
+		http.Error(w, "Internal server error", http.StatusInternalServerError)
+		return
+	}
+	var activeKeys, revokedKeys []pages.AccessKeyRow
+	for _, k := range keys {
+		row := buildAccessKeyRow(k)
+		if k.RevokedAt != nil {
+			revokedKeys = append(revokedKeys, row)
+		} else {
+			activeKeys = append(activeKeys, row)
+		}
+	}
+	var activeClients, revokedClients []pages.AccessClientRow
+	for _, c := range clients {
+		row := buildAccessClientRow(c)
+		if c.RevokedAt != nil {
+			revokedClients = append(revokedClients, row)
+		} else {
+			activeClients = append(activeClients, row)
+		}
+	}
+	data := BaseTemplateData(r, sm, "api-keys", "Access")
+	props := pages.AccessProps{
+		IsAdmin:           IsAdmin(sm, r),
+		CSRFToken:         GetCSRFToken(r),
+		ActiveKeys:        activeKeys,
+		RevokedKeys:       revokedKeys,
+		HasAnyKeys:        len(keys) > 0,
+		ActiveClients:     activeClients,
+		RevokedClients:    revokedClients,
+		HasAnyClients:     len(clients) > 0,
+		JustCreatedKey:    keyReveal,
+		JustCreatedClient: clientReveal,
+	}
 	renderSettingsTab(tr, w, r, data, pages.SettingsTabAccess, pages.Access(props))
+}
+
+// popAPIKeyReveal consumes the one-time plaintext flash an API-key create
+// left behind (no-JS path only), so the Access tab can render the copy-now
+// block inline on the redirect-back render. Returns nil when no key was
+// just created.
+func popAPIKeyReveal(r *http.Request, sm *scs.SessionManager) *pages.AccessReveal {
+	secret := sm.PopString(r.Context(), "created_api_key")
+	if secret == "" {
+		return nil
+	}
+	return &pages.AccessReveal{
+		Name:   sm.PopString(r.Context(), "created_api_key_name"),
+		Secret: secret,
+		Scope:  sm.PopString(r.Context(), "created_api_key_scope"),
+	}
 }
 
 // buildAccessKeyRow flattens a service.APIKeyResponse into the templ-side
@@ -148,78 +170,44 @@ func buildAccessClientRow(c service.OAuthClientResponse) pages.AccessClientRow {
 	}
 }
 
-// APIKeyNewPageHandler serves GET /admin/api-keys/new.
-func APIKeyNewPageHandler(sm *scs.SessionManager, tr *TemplateRenderer) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		data := BaseTemplateData(r, sm, "api-keys","Create API Key")
-		renderAPIKeyNew(w, r, tr, data, pages.APIKeyNewProps{
-			CSRFToken: GetCSRFToken(r),
-			Breadcrumbs: []components.Breadcrumb{
-				{Label: "API Keys", Href: "/settings/api-keys"},
-				{Label: "Create API Key"},
-			},
-		})
-	}
-}
-
-// renderAPIKeyNew hosts the typed APIKeyNewProps inside the Settings
-// shell as a sub-view of the API Keys tab so the rail stays visible.
-func renderAPIKeyNew(w http.ResponseWriter, r *http.Request, tr *TemplateRenderer, data map[string]any, props pages.APIKeyNewProps) {
-	renderSettingsTab(tr, w, r, data, pages.SettingsTabAccess, pages.APIKeyNew(props))
-}
-
-// APIKeyCreatePageHandler serves POST /admin/api-keys/new.
-// Creates the key and redirects to the "created" page that shows the plaintext key once.
+// APIKeyCreatePageHandler serves POST /settings/api-keys/new.
+//
+// One-click create: the name is optional (defaults to "API key") and the
+// scope defaults to full access, so the bare button mints immediately. The
+// plaintext is stashed in a one-shot session flash and we 303 back to the
+// Access tab, where popAPIKeyReveal renders it inline — no /created
+// subpage. The in-page swapper follows the redirect as a fragment so the
+// reveal lands without a full navigation.
 func APIKeyCreatePageHandler(svc *service.Service, sm *scs.SessionManager, tr *TemplateRenderer) http.HandlerFunc {
 	return func(w http.ResponseWriter, r *http.Request) {
 		name := strings.TrimSpace(r.FormValue("name"))
 		if name == "" {
-			FlashRedirect(w, r, sm, "error", "Name is required", "/settings/api-keys/new")
-			return
+			name = "API key"
 		}
 		scope := r.FormValue("scope")
-		if scope == "" {
+		if scope != "read_only" {
 			scope = "full_access"
 		}
 		result, err := svc.CreateAPIKeyLegacy(r.Context(), name, scope)
 		if err != nil {
-			FlashRedirect(w, r, sm, "error", "Failed to create API key", "/settings/api-keys/new")
+			FlashRedirect(w, r, sm, "error", "Failed to create API key", "/settings/api-keys")
 			return
 		}
-		// Store the plaintext key in the session so the "created" page can display it once.
+		reveal := &pages.AccessReveal{Name: result.Name, Secret: result.PlaintextKey, Scope: scope}
+		// JS path: the in-page swapper POSTs with the fragment header.
+		// Render the Access tab fragment directly with the reveal inline —
+		// no redirect, no one-shot flash to race against.
+		if r.Header.Get(settingsFragmentHeader) == "1" {
+			renderAccessTab(svc, sm, tr, w, r, reveal, nil)
+			return
+		}
+		// No-JS fallback: stash the reveal in a flash and 303 back so a
+		// form post doesn't re-submit on refresh.
 		sm.Put(r.Context(), "created_api_key", result.PlaintextKey)
 		sm.Put(r.Context(), "created_api_key_name", result.Name)
-		http.Redirect(w, r, "/settings/api-keys/"+result.ID+"/created", http.StatusSeeOther)
+		sm.Put(r.Context(), "created_api_key_scope", scope)
+		http.Redirect(w, r, "/settings/api-keys", http.StatusSeeOther)
 	}
-}
-
-// APIKeyCreatedPageHandler serves GET /admin/api-keys/{id}/created.
-// Shows the plaintext key ONE TIME (from session flash).
-func APIKeyCreatedPageHandler(sm *scs.SessionManager, tr *TemplateRenderer) http.HandlerFunc {
-	return func(w http.ResponseWriter, r *http.Request) {
-		key := sm.PopString(r.Context(), "created_api_key")
-		name := sm.PopString(r.Context(), "created_api_key_name")
-		if key == "" {
-			// Key already shown or session expired — redirect to list.
-			http.Redirect(w, r, "/settings/api-keys", http.StatusSeeOther)
-			return
-		}
-		data := BaseTemplateData(r, sm, "api-keys","API Key Created")
-		renderAPIKeyCreated(w, r, tr, data, pages.APIKeyCreatedProps{
-			KeyName:      name,
-			PlaintextKey: key,
-			Breadcrumbs: []components.Breadcrumb{
-				{Label: "API Keys", Href: "/settings/api-keys"},
-				{Label: "Key Created"},
-			},
-		})
-	}
-}
-
-// renderAPIKeyCreated hosts the typed APIKeyCreatedProps inside the
-// Settings shell as a sub-view of the API Keys tab.
-func renderAPIKeyCreated(w http.ResponseWriter, r *http.Request, tr *TemplateRenderer, data map[string]any, props pages.APIKeyCreatedProps) {
-	renderSettingsTab(tr, w, r, data, pages.SettingsTabAccess, pages.APIKeyCreated(props))
 }
 
 // APIKeyRevokePageHandler serves POST /admin/api-keys/{id}/revoke.
