--- internal/templates/components/pages/api_key_created.templ
@@ -1,48 +1,0 @@
-package pages
-
-import "breadbox/internal/templates/components"
-
-// APIKeyCreated renders the /settings/api-keys/{id}/created page — shown once after
-// creating a new API key, displaying the plaintext key for a one-time copy.
-// Hosts inside the base layout via TemplateRenderer.RenderWithTempl. The
-// markup mirrors the previous html/template version byte-for-byte: same
-// Alpine clipboard binding, same revealed-once flow, same "Done" link.
-templ APIKeyCreated(p APIKeyCreatedProps) {
-	@components.BreadcrumbNav(p.Breadcrumbs)
-	<div class="bb-page-header">
-		<div>
-			<h1 class="bb-page-title">API Key Created</h1>
-			<p class="text-sm text-base-content/50 mt-1">Copy your key now -- you won't see it again</p>
-		</div>
-	</div>
-	<div class="bb-card p-6 max-w-lg">
-		<div class="flex items-center gap-3 mb-5">
-			<div class="w-9 h-9 rounded-lg bg-success/10 flex items-center justify-center">
-				@components.LucideIcon("check", "w-5 h-5 text-success")
-			</div>
-			<div>
-				<h2 class="text-sm font-semibold">{ p.KeyName }</h2>
-				<p class="text-xs text-base-content/45">Copy your key now -- it won't be shown again.</p>
-			</div>
-		</div>
-		<div class="mb-5">
-			<label class="text-xs font-medium text-base-content/50 mb-1.5 block">API Key</label>
-			<div class="flex flex-col sm:flex-row gap-2" x-data="{ copied: false }">
-				<div class="relative flex-1 min-w-0">
-					<input type="text" value={ p.PlaintextKey } readonly id="api-key-value" class="input input-sm font-mono w-full bg-base-200/50 border-base-300 text-xs truncate"/>
-				</div>
-				<button class="btn btn-primary btn-sm shrink-0 gap-2 w-full sm:w-auto" :class="copied && '!btn-success'" @click="navigator.clipboard.writeText(document.getElementById('api-key-value').value).then(function(){ $data.copied = true; setTimeout(function(){ $data.copied = false; }, 2000); })">
-					<i data-lucide="clipboard" class="w-4 h-4" x-show="!copied"></i>
-					<i data-lucide="check" class="w-4 h-4" x-show="copied" x-cloak></i>
-					<span x-text="copied ? 'Copied!' : 'Copy'"></span>
-				</button>
-			</div>
-		</div>
-		<div class="flex flex-wrap items-center justify-between gap-2 pt-4 border-t border-base-300">
-			<p class="text-xs text-base-content/40 min-w-0">
-				Use in the <code class="font-mono text-primary/70 bg-primary/5 px-1.5 py-0.5 rounded">X-API-Key</code> header for API requests.
-			</p>
-			<a href="/settings/api-keys" class="btn btn-ghost btn-sm shrink-0">Done</a>
-		</div>
-	</div>
-}
