--- internal/templates/components/pages/oauth_client_new.templ
@@ -1,44 +1,0 @@
-package pages
-
-import "breadbox/internal/templates/components"
-
-// OAuthClientNew renders the /oauth-clients/new page — a small form that
-// creates a new OAuth client (name + scope) for external connectors such
-// as Claude. Hosts inside the base layout via
-// TemplateRenderer.RenderWithTempl. The markup mirrors the previous
-// html/template version byte-for-byte; the form posts to
-// /oauth-clients/new which is handled by OAuthClientCreatePageHandler in
-// admin/oauth.go.
-templ OAuthClientNew(p OAuthClientNewProps) {
-	@components.BreadcrumbNav(p.Breadcrumbs)
-	<div class="bb-page-header">
-		<div>
-			<h1 class="bb-page-title">Create OAuth Client</h1>
-			<p class="text-sm text-base-content/50 mt-1">Generate credentials for an external connector (e.g., Claude)</p>
-		</div>
-	</div>
-	<div class="bb-card p-6 max-w-lg">
-		<form method="POST" action="/oauth-clients/new">
-			<input type="hidden" name="_csrf" value={ p.CSRFToken }/>
-			<fieldset class="fieldset mb-5">
-				<label class="text-sm font-medium text-base-content/70 mb-1.5 block" for="name">Name</label>
-				<input type="text" id="name" name="name" class="input w-full" placeholder="e.g., Claude Connector, Scheduled Agent" required autofocus/>
-				<p class="text-xs text-base-content/40 mt-2">A descriptive name shown on the consent screen when authorizing.</p>
-			</fieldset>
-			<fieldset class="fieldset mb-6">
-				<label class="text-sm font-medium text-base-content/70 mb-1.5 block" for="scope">Scope</label>
-				<select id="scope" name="scope" class="select w-full">
-					<option value="full_access">Full Access — read and write via MCP</option>
-					<option value="read_only">Read Only — query data only</option>
-				</select>
-				<p class="text-xs text-base-content/40 mt-2">Controls what the connected service can do. Full access allows transaction categorization, rule management, and write MCP tools.</p>
-			</fieldset>
-			<div class="flex gap-3 pt-2">
-				<button type="submit" class="btn btn-primary btn-sm">
-					@components.LucideIcon("plus", "w-4 h-4") Create OAuth Client
-				</button>
-				<a href="/access" class="btn btn-ghost btn-sm">Cancel</a>
-			</div>
-		</form>
-	</div>
-}
