--- internal/templates/components/pages/oauth_client_new_types.go
@@ -1,13 +1,0 @@
-//go:build !headless && !lite
-
-package pages
-
-import "breadbox/internal/templates/components"
-
-// OAuthClientNewProps mirrors the data map the old oauth_client_new.html
-// read off the layout's data map. The handler resolves the breadcrumb
-// trail and CSRF token before calling the templ component.
-type OAuthClientNewProps struct {
-	Breadcrumbs []components.Breadcrumb
-	CSRFToken   string
-}
