--- internal/templates/components/pages/settings_host.templ
@@ -1,25 +1,0 @@
-package pages
-
-
-import "breadbox/internal/templates/components"
-// SettingsHost is the "background page" rendered when a /settings/* URL
-// is hit directly (cold deep-link). The modal opens on top of it via the
-// SettingsModal mounted globally in base.html — this just provides the
-// shell so the sidebar + topbar render normally.
-//
-// The visible content is intentionally minimal: a faint heading so the
-// page isn't blank if the modal fails to open for some reason
-// (no-script, blocked dialog API). Once the user closes the modal, the
-// Alpine factory navigates them home — they won't dwell on this view.
-templ SettingsHost() {
-	<div class="max-w-2xl mx-auto py-12 text-center text-base-content/40">
-		@components.LucideIcon("settings", "w-8 h-8 mx-auto mb-3 opacity-60")
-		<p class="text-sm">Loading settings…</p>
-		<noscript>
-			<p class="text-xs mt-2">
-				Settings requires JavaScript. Enable it and reload, or
-				visit the underlying pages directly via the sidebar.
-			</p>
-		</noscript>
-	</div>
-}
